+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 5: Object Oriented JavaScript - ES5 & ES2015
     - Constructors & the 'this' Keyword
     - Built In Constructors
+    - Prototypes Explained
 - An example with constructors and the this keyword was added.
 - An example with built-in constructors was added.
+- An example with prototypes was added.
 
 ## August 27, 2020
 - Modern JavaScript From The Beginning (Udemy)